[compiler]add new feature to remove-copy pass. (#141)
The current cases are coming to handle the case where srcAlloc and
target are of the same type or targetAlloc and src are of the same type,
I added the case where srcAlloc and target are of different types. See
the test case for details.Feel free to leave suggestions below.