[GEMM codegen] Distribute Shared memory copy (#303)
Precondition: Check whether the computation can be vectorized. If it
cannot, fall back to the non-vectorized algorithm.

1. Convert linalg.copy to linalg.generic.

2. Perform tiling. Total number of elements = N. The number of elements
that can be read in one vectorized operation = V. So the number of loops
equals N divided by V times the number of threads. In the current case,
this is
8×128. Once the thread block completes a tile, it moves to the next
tile.
3. Within the tile, distribute by introducing threadIdx.x and
threadIdx.y, calculate the flattened 1D coordinate, and then distribute,
each thread should copy a "1x8" tile.
4. Transform vectorized operations to transfer_read and transfer_write.
5. Outer loop unroll.