[compiler] fix DeviceGraphCluster for if op (#362)
1. support nested region such as scf.if for GraphClusteringByDevice pass
2. scf.if output type compatibility
a. do not insert shape_ext.TieOp after a mhlo.ConvertOp when Convert
static shape op to dynamic shape op
b. set return type for thenBlock and elseBlock for scf.ifOp when the
shape of the return value of scf.ifOp can be deduced
         by shape_ext.tieOp
c. set bounded shape for thenBlock and elseBlok of scf.ifOp to meet the
compatibility requirements of scf.ifop return
        types
3. decoupling static shape analysis from bounded shape analysis